Login system added

+++ Web_Files/model/data_encode.php
@@ -1,22 +1,54 @@
 <?php
 
-function addUser($userData){
+function addUser($userData)
+{
 
-    if ($userData['password'] == $userData['passwordConfirm']){
-        $fullUser = array(
-            "email" => $userData['email'],
-            "password" => $userData['password']
-        );
+    if ($userData['password'] == $userData['passwordConfirm']) {
 
         $data = file_get_contents("model/data/dataTest.json");
-        $data = json_decode($data,JSON_OBJECT_AS_ARRAY);
+        $data = json_decode($data, JSON_OBJECT_AS_ARRAY);
+        $userAlreadyExist = 0;
 
-        $data = (!empty($data))?$data:[];
+        foreach ($data as $row) {
+            foreach ($row as $roow) {
+                if ($roow == $userData['email']) {
+                    $userAlreadyExist = $userAlreadyExist + 1;
+                } else {
+                    $userAlreadyExist = $userAlreadyExist + 0;
+                }
+            }
+        }
 
-        array_push($data,$fullUser);
-        $data = json_encode($data, JSON_PRETTY_PRINT);
+        if ($userAlreadyExist == 0) {
+            echo "new user ";
+            $fullUser = array(
+                "email" => $userData['email'],
+                "password" => $userData['password']
+            );
 
-        file_put_contents("model/data/dataTest.json", $data);
-        require "view/home.php";
+            $data = file_get_contents("model/data/dataTest.json");
+            $data = json_decode($data, JSON_OBJECT_AS_ARRAY);
+
+            $data = (!empty($data)) ? $data : [];
+
+            array_push($data, $fullUser);
+            $data = json_encode($data, JSON_PRETTY_PRINT);
+
+            file_put_contents("model/data/dataTest.json", $data);
+            echo $userData['email'];
+            if(session_start()) {
+                $_SESSION['email'] = $userData['email'];
+            }
+            header("Location: /home");
+            require "view/home.php";
+        } else {
+            require "view/register.php";
+            echo "Email déjà existante !";
+        }
+
+    } else {
+        require "view/register.php";
+        echo "Mot de passe incorrect !";
     }
+
 }
